+++ R/utils_cbh.R
@@ -1152,27 +1152,44 @@ clean_cbh_df <- function(cbh_df = NULL, trees_poly, lad_profile, force_cbh_lte_h
     nrow(cbh_df)>0
   ){
     # filter based on cbh vs ht
-    cbh_df <- cbh_df %>%
-      dplyr::inner_join(
-        trees_poly %>%
-          sf::st_drop_geometry() %>%
-          dplyr::select(treeID,tree_height_m)
-        , by = "treeID"
-      ) %>%
+    cbh_df <- trees_poly %>%
+      dplyr::select(treeID,tree_height_m) %>%
+      dplyr::inner_join(cbh_df, by = "treeID") %>%
       dplyr::filter(
         !is.na(tree_cbh_m)
         & tree_cbh_m < tree_height_m
       ) %>%
-      dplyr::mutate(is_training_cbh=T)
+      dplyr::mutate(is_training_cbh=T) %>%
+      make_spatial_predictors()
   }else if(nrow(cbh_df)>0){
     # filter cbh
-    cbh_df <- cbh_df %>%
+    cbh_df <- trees_poly %>%
+      dplyr::select(treeID) %>%
+      dplyr::inner_join(cbh_df, by = "treeID") %>%
       dplyr::filter(
         !is.na(tree_cbh_m)
       ) %>%
-      dplyr::mutate(is_training_cbh=T)
+      dplyr::mutate(is_training_cbh=T) %>%
+      make_spatial_predictors()
   }
 
   # return
   return(cbh_df)
+}
+
+################################################################################################################################################
+## function to make spatial predictor variables from POLYGON sf data with treeID and tree_height_m
+# prep data for missing data model by creating predictor vars suffixed "_zzz"
+################################################################################################################################################
+make_spatial_predictors <- function(poly_sf) {
+  df <- poly_sf %>%
+    # prep data for missing data model by creating predictor vars suffixed "_zzz"
+    dplyr::mutate(crown_area_zzz = sf::st_area(.) %>% as.numeric()) %>%
+    sf::st_centroid() %>%
+    dplyr::mutate(
+      tree_x_zzz = sf::st_coordinates(.)[,1]
+      , tree_y_zzz = sf::st_coordinates(.)[,2]
+    ) %>%
+    sf::st_drop_geometry()
+  return(df)
 }